Restructure components and code review suggestions
