+++ components/timesheet-entry/timesheet-entry.scss
@@ -1,4 +1,4 @@
-@import "../../../../shared/styles/main";
+@import "../../shared/styles/main";
 
 .timesheet-entry-wrapper {
   border: solid 1px $color-accent-2;
