+++ components/timesheet-input/timesheet-input.scss
@@ -1,4 +1,4 @@
-@import "../../../shared/styles/main";
+@import "../../shared/styles/main";
 
 .timesheet-field {
   background-color: $color-accent-2;
