+++ components/timesheet/timesheet.scss
@@ -1,8 +1,7 @@
-@import "../../../shared/styles/main";
+@import "../../shared/styles/main";
 
 .timesheet-wrapper {
   background-color: $color-monochrome-1;
-  // background-color: $color-accent-2;
   margin: 20px auto 0;
   width: 90%;
