+++ pages/index.scss
@@ -1,4 +1,4 @@
-@import "../../shared/styles/main";
+@import "../shared/styles/main";
 
 .timesheet-page-wrapper {
   display: flex;
